+++ extensions/broker/src/main/java/de/sovity/extension/broker/service/IdsBrokerServiceImpl.java
@@ -20,6 +20,11 @@ import de.sovity.extension.broker.sender.message.UnregisterConnectorMessage;
 import de.sovity.extension.broker.sender.message.UnregisterResourceMessage;
 import org.eclipse.edc.connector.contract.spi.offer.store.ContractDefinitionStore;
 import org.eclipse.edc.connector.contract.spi.types.offer.ContractDefinition;
+import org.eclipse.edc.connector.policy.spi.PolicyDefinition;
+import org.eclipse.edc.connector.policy.spi.store.PolicyDefinitionStore;
+import org.eclipse.edc.policy.model.AtomicConstraint;
+import org.eclipse.edc.policy.model.LiteralExpression;
+import org.eclipse.edc.policy.model.Rule;
 import org.eclipse.edc.protocol.ids.service.ConnectorServiceSettings;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.asset.AssetIndex;
@@ -55,9 +60,13 @@ public class IdsBrokerServiceImpl implements IdsBrokerService, EventSubscriber {
 
     private final ContractDefinitionStore contractDefinitionStore;
 
+    private final PolicyDefinitionStore policyDefinitionStore;
+
     private final URL brokerBaseUrl;
 
     private final AssetIndex assetIndex;
+
+    private final List<String> policyBrokerBlacklist;
 
     private final Monitor monitor;
 
@@ -66,15 +75,19 @@ public class IdsBrokerServiceImpl implements IdsBrokerService, EventSubscriber {
             ConnectorServiceSettings connectorServiceSettings,
             Hostname hostname,
             ContractDefinitionStore contractDefinitionStore,
+            PolicyDefinitionStore policyDefinitionStore,
             URL brokerBaseUrl,
             AssetIndex assetIndex,
+            List<String> policyBrokerBlacklist,
             Monitor monitor) {
         this.dispatcherRegistry = dispatcherRegistry;
         this.connectorServiceSettings = connectorServiceSettings;
         this.hostname = hostname;
         this.contractDefinitionStore = contractDefinitionStore;
+        this.policyDefinitionStore = policyDefinitionStore;
         this.brokerBaseUrl = brokerBaseUrl;
         this.assetIndex = assetIndex;
+        this.policyBrokerBlacklist = policyBrokerBlacklist;
         this.monitor = monitor;
     }
 
@@ -229,15 +242,51 @@ public class IdsBrokerServiceImpl implements IdsBrokerService, EventSubscriber {
         var contractDefinition = contractDefinitionStore.findById(contractDefinitionId);
 
         var resourceMap = new HashMap<String, Asset>();
-        var assetsFromContractDefinition = getAssetsFromContractDefinition(contractDefinition);
-        for (var asset : assetsFromContractDefinition) {
-            var resourceId = String.format("%s-%s",
-                    contractDefinition.getId(),
-                    asset.getId());
-            resourceMap.put(resourceId, asset);
+        var hasBrokerBlacklistedPolicy = checkForBlacklistedPolicy(contractDefinition);
+
+        if (hasBrokerBlacklistedPolicy) {
+            monitor.info("Not publishing resource at broker due to possible policy breach.");
+        } else {
+            var assetsFromContractDefinition = getAssetsFromContractDefinition(contractDefinition);
+            for (var asset : assetsFromContractDefinition) {
+                var resourceId = String.format("%s-%s",
+                        contractDefinition.getId(),
+                        asset.getId());
+                resourceMap.put(resourceId, asset);
+            }
         }
 
         return resourceMap;
+    }
+
+    private boolean checkForBlacklistedPolicy(ContractDefinition contractDefinition) {
+        var contractPolicyId = contractDefinition.getContractPolicyId();
+        var accessPolicyId = contractDefinition.getAccessPolicyId();
+
+        var contractConstraints = getConstraints(policyDefinitionStore.findById(contractPolicyId));
+        var accessConstraints = getConstraints(policyDefinitionStore.findById(accessPolicyId));
+        return containsBlacklistedPolicy(contractConstraints) ||
+                containsBlacklistedPolicy(accessConstraints);
+    }
+
+    private List<AtomicConstraint> getConstraints(PolicyDefinition contractPolicy) {
+        return contractPolicy
+                .getPolicy()
+                .getPermissions()
+                .stream()
+                .map(Rule::getConstraints)
+                .flatMap(List::stream)
+                .map(AtomicConstraint.class::cast)
+                .toList();
+    }
+
+    private boolean containsBlacklistedPolicy(List<AtomicConstraint> constraints) {
+        return constraints
+                .stream()
+                .map(AtomicConstraint::getLeftExpression)
+                .map(LiteralExpression.class::cast)
+                .map(LiteralExpression::getValue)
+                .anyMatch(policyBrokerBlacklist::contains);
     }
 
     private List<Asset> getAssetsFromContractDefinition(ContractDefinition contractDefinition) {