+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [X.X.X]
+### Feature:
+- Add setting `POLICY_BROKER_BLACKLIST` to blacklist policies from being published to broker
 
 ## [1.1.0] - 2023-01-23